fix: changed migration & add local & change API & setup Swagger

+++ src/main/resources/db/migration/V4__create_doctors_table.sql
@@ -3,6 +3,6 @@ CREATE TABLE doctors (
     doctor_specialization VARCHAR(120) NOT NULL,
     doctor_experience INTEGER NOT NULL,
     doctor_qualification VARCHAR(120) NOT NULL,
-    doctor_user_id INTEGER REFERENCES users(user_id),
-    doctor_organization INTEGER REFERENCES medical_organizations(organization_id)
+    doctor_user_id BIGINT REFERENCES users(user_id),
+    doctor_organization BIGINT REFERENCES medical_organizations(organization_id)
 )
