+++ src/main/resources/db/migration/V7__create_medical_card_table.sql
@@ -1,5 +1,5 @@
 CREATE TABLE "medical_card" (
     card_id BIGSERIAL PRIMARY KEY,
-    record_id INTEGER REFERENCES medical_record(record_id),
-    patient_id INTEGER REFERENCES patients(patient_id)
+    record_id BIGINT REFERENCES medical_record(record_id),
+    patient_id BIGINT REFERENCES patients(patient_id)
 )
